Move isolated FG23 worker classes to the java8 sourceset
