+++ plugin/src/java8/java/com/gtnewhorizons/retrofuturagradle/java8/fg23/ArtifactSaver.java
@@ -8,7 +8,7 @@
  * this library; if not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
  * 02110-1301 USA
  */
-package com.gtnewhorizons.retrofuturagradle.mcp.fg23;
+package com.gtnewhorizons.retrofuturagradle.java8.fg23;
 
 import java.io.File;
 import java.io.FileOutputStream;
